rework hw-2.15
